+++ NEWS.md
@@ -9,6 +9,7 @@ these were part of the htmlwidget which is useless.
 - Initial `tippy_plot` to use plots in tooltips.
 - Support new themes.
 - Allow calling methods on tooltips (e.g.: disable, show, hide, etc.)
+- Allow setting global theme, see [#3](https://github.com/JohnCoene/tippy/issues/3)
 
 ## tippy 0.1.0